+++ neo/renderer/glsl/zfill.frag
@@ -1,10 +1,10 @@
 #version 100
 
-precision mediump float;
+precision highp float;
 
 uniform sampler2D u_fragmentMap0;
-uniform lowp float u_alphaTest;
-uniform lowp vec4 u_glColor;
+uniform mediump float u_alphaTest;
+uniform mediump vec4 u_glColor;
 
 varying vec2 var_TexDiffuse;
